+++ emergence_lib/src/structures/ghost.rs
@@ -7,7 +7,7 @@
 use bevy::prelude::*;
 
 use crate::{
-    player_interaction::clipboard::StructureData,
+    player_interaction::{clipboard::StructureData, selection::ObjectInteraction},
     simulation::geometry::{Facing, TilePos},
 };
 
@@ -17,11 +17,17 @@ use super::{crafting::InputInventory, StructureId};
 #[derive(Component, Clone, Copy, Debug)]
 pub(crate) struct Ghost;
 
+/// A marker component indicating that this structure should be rendered in a transparent style.
+#[derive(Component, Clone, Copy, Debug)]
+pub(crate) struct Ghostly;
+
 /// The set of components needed to spawn a ghost.
 #[derive(Bundle)]
 pub(super) struct GhostBundle {
     /// Marker component
     ghost: Ghost,
+    /// Render this entity in a translucent style
+    ghostly: Ghostly,
     /// The location of the ghost
     tile_pos: TilePos,
     /// The variety of structure
@@ -30,6 +36,8 @@ pub(super) struct GhostBundle {
     facing: Facing,
     /// The items required to actually seed this item
     construction_materials: InputInventory,
+    /// How is this structure being interacted with
+    object_interaction: ObjectInteraction,
 }
 
 impl GhostBundle {
@@ -41,10 +49,12 @@ impl GhostBundle {
     ) -> Self {
         GhostBundle {
             ghost: Ghost,
+            ghostly: Ghostly,
             tile_pos,
             structure_id: data.structure_id,
             facing: data.facing,
             construction_materials,
+            object_interaction: ObjectInteraction::None,
         }
     }
 }
@@ -58,12 +68,16 @@ pub(crate) struct Preview;
 pub(super) struct PreviewBundle {
     /// Marker component
     preview: Preview,
-    /// The location of the ghost
+    /// Render this entity in a translucent style
+    ghostly: Ghostly,
+    /// The location of the preview
     tile_pos: TilePos,
     /// The variety of structure
     structure_id: StructureId,
-    /// The direction the ghost is facing
+    /// The direction the preview is facing
     facing: Facing,
+    /// How is this structure being interacted with
+    object_interaction: ObjectInteraction,
 }
 
 impl PreviewBundle {
@@ -71,9 +85,11 @@ impl PreviewBundle {
     pub(super) fn new(tile_pos: TilePos, data: StructureData) -> Self {
         PreviewBundle {
             preview: Preview,
+            ghostly: Ghostly,
             tile_pos,
             structure_id: data.structure_id,
             facing: data.facing,
+            object_interaction: ObjectInteraction::Hovered,
         }
     }
 }
